Database completed.

+++ Entities/Concrete/Car.cs
@@ -15,7 +15,7 @@ namespace Entities.Concrete
 
         public string ModelYear { get; set; }
 
-        public int DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
 
         public string Description { get; set; }
